pyproject.toml
[build-system]
requires      = ["setuptools>=45"]
build-backend = "setuptools.build_meta"


[project]
name = "BPD"
authors = [{ name = "Ismael Mendoza" }]
description = "Bayesian Pixel Domain method for shear inference."
version = "0.0.1"
license = { file = "LICENSE" }
readme = "README.md"
dependencies = [
    "numpy >=2.0.0",
    "galsim >=2.3.0",
    "blackjax >=1.2.0",
    "optax >=0.2.0",
    "numpyro >=0.13.0",
    "quadax>=0.2.5",
    "tensorflow-probability >=0.21.0",
    "arviz>=0.21.0",
]


[project.optional-dependencies]
dev = [
    "pytest",
    "typer",
    "ruff",
    "ChainConsumer>=1.2.0", #v1.2.2 still requires python < 3.13
    "ipython",
    "ipykernel",
]


[project.urls]
home = "https://github.com/LSSTDESC/BPD"


[tool.setuptools.packages.find]
include = ["bpd*"]

[tool.ruff]
exclude = [
    ".bzr",
    ".direnv",
    ".eggs",
    ".git",
    ".git-rewrite",
    ".hg",
    ".ipynb_checkpoints",
    ".mypy_cache",
    ".nox",
    ".pants.d",
    ".pyenv",
    ".pytest_cache",
    ".pytype",
    ".ruff_cache",
    ".svn",
    ".tox",
    ".venv",
    ".vscode",
    "__pypackages__",
    "_build",
    "buck-out",
    "build",
    "dist",
    "node_modules",
    "site-packages",
    "venv",
]

# Same as Black.
line-length  = 88
indent-width = 4

# Assume Python 3.12
target-version = "py312"

[tool.ruff.format]
# Like Black, use double quotes for strings.
quote-style = "double"

# Like Black, indent with spaces, rather than tabs.
indent-style = "space"

# Like Black, respect magic trailing commas.
skip-magic-trailing-comma = false

# Like Black, automatically detect the appropriate line ending.
line-ending = "auto"

exclude = ["*.ipynb"]


[tool.ruff.lint]
select = ["E", "F", "I", "W", "B", "SIM", "PLE", "PLC", "PLW", "RUF"]
ignore = [
    "C901",    # I want to decide what is 'too' complex
    "E501",    # I don't care about formatting docstrings
    "E731",    # I like lambda's sometimes
    "PLC0206", # too picky, not always useful
    "RUF027",  # does not account for shell scripts correctly
    "RUF052",  # I find underscore variables useful
]
preview = true
exclude = ["*.ipynb", "scripts/one_galaxy_shear.py", "scripts/benchmarks/*.py"]


[tool.pytest.ini_options]
minversion     = "6.0"
addopts        = "-ra -v --strict-markers"
filterwarnings = ["ignore::DeprecationWarning:tensorflow.*"]
markers        = ["slow: marks tests as slow (deselect with '-m \"not slow\"')"]
